feat: create a hook for customizing `sonner`'s toaster
Create `src/hook/useToast.tsx` hook for customizing the
toaster of `sonner` package and providing a custom API
to fire toasters.

The customization for now is to paint the icon of `sonner`'s
toaster based of the type or the variant of the toaster
whether is "error", "info, "success" or "warning".