types: create a file to define global types in the app
Create `src/types/global.ts` file and define inside it four
types represent some objects in the application.

The four times are:

- `ResourceDateInfo` type which represents two date
information, when the resource had created and when
was the last modification on the resource.

- `PaginationModel` type which represents the pagination
model when fetching some data page by page.

- `FetchError` class which can de used as type, this class is
a custom error object represent the error object that
supposed to come from the server.

- `ErrorDetails` type which represents the object that
contains the error details inside instance of `FetchError`
class.