refactor: separate `fetchAPI` function's logic to 2 files
Separate the section that handles preparing the headers,
url path and search params of the request,  And sends
the request using the native fetch function, to a server
action function called `customFetchAPI` in
`src/lib/customFetchAPI.ts` file.

And convert `fetchAPI` function from a server action to a
function works normally in the client, calls `customFetchAPI`
function and handles returning the response and throwing
an error if the request failed.