feat: create `useRecentInstructionsActivitiesState` hook
Create `useRecentInstructionsActivitiesState` hook to manage
the state of the instructions in the fetched data in
`useRecentActivities` hook, This hook provides three functions:

- `addToRecentInstructionsActivities` to add new instructions
to the state of the fetched recent activities data.

- `updateRecentInstructionsActivities` to update the state of
instructions in the state of the fetched recent activities data.

- `removeRecentInstructionsActivities` to remove instructions
from the state of the fetched recent activities data.