feat: create `useInstructionsStatesManagement` hook
Create `useInstructionsStatesManagement` hook for collecting
all hooks that update datasets instructions states in one place
and providing methods to handle all related states updates
together.

This hook provides three methods:

- `addNewInstructionState` to call all states updates functions
that are related to adding a new instruction together.

- `updateInstructionState` to call all states updates functions
that are related to updating an instruction together.

- `deleteInstructionState` to call all states updates functions
that are related to deleting an instruction together.