feat: create dataset page
Create dataset page where a single dataset being fully
displayed with its instructions and the forms of adding
and updating its instructions.

Wrap the content of the page with the instructions context
provider component `DatasetInstructionsContext` that
creates and shares states for the instructions of the opened
dataset in the page.

For now the UI of the page just contins the card that
displays the full details of the opened dataset in the
page.