feat: create `startExportingDataset` function
First create the types of datasets formats and files extentions
in `src/types/datasets.ts`.

Then create `startExportingDataset` function which is the
function that starts exporting dataset by a few steps:

1. taking the id of the dataset and the format that should
the dataset be exported in.

2. requesting saving file from the user and creating the
writable stream to the chosen file by the user.

3. prepareing API endpoint and creating a new web socket
connection.

4. And in the end returning the web socket instance and
the writable stream to the chosen file by the user.