feat: create `DatasetExportingState` class
Create `DatasetExportingState` class which each instance
of this class represent an "exporting dataset state",

When exporting a dataset requested, an instance of this
class will be created to represent and manage the state of
this new exporting process in the app,

The instance handles starting the exporting process, tracking
the progress of exporting process, canceling exporting process
and resetting the state of exporting process after exporting
stoped for whatever reason.

Also, Because React cannot track the state of the instances
of `DatasetExportingState` class when any updates occurr
to them and rerender the UI component,

Add `updator` method to the class which is supposed to be a
function triggers an effect causes React to rerender the UI,
this method should be called when any updates occurr inside
the instances of `DatasetExportingState` class so that React
can know that there is an update occurred and the UI
component should be rerendered.