feat: create `useDatasetsExportingStatesRegistery` hook
Create `useDatasetsExportingStatesRegistery` hook which is
the place that the instances of `DatasetExportingState` class
should be stored in,

This hook uses `useState` hook with `Map` object to store the
exporting states of the datasets while the ids of the datasets
are the keys and the mapped values to these keys are instances
of `DatasetExportingState` class containing the states of these
datasets ids.

In fact, It is `useMap` hook from `https://usehooks-ts.com/`
but modified a little.