feat: create `ExportingDatasetsContext` component
Create `ExportingDatasetsContext` component to create a
context to share the states of exporting datasets and utils
functions to provide the ability to export dataset from anywhere
in the app, and to provide get and remove datasets exporting states.

Wrap the root layout with `ExportingDatasetsContext`
component to all components in te app be able access
this context via `useContext` hook.

Create `useExportingDatasetsContext` to provide a direct way
to access exporting datasets context instead of importing
`useContext` hook and `exportingDatasetsContext` context
in all places.