feat: create `DatasetExportingState` component
Create `DatasetExportingState` component to display
the exporting state of the opened dataset in dataset
page,

This component displays in the UI the states of dataset
exporting (if the dataset is being exported) like exporting
progress and exporting state if is canceled, done successfully
or failed to complate.

And also the component provides an option to cancel the
exporting process.