feat: don't remove instructions activities & register a new instead
Remove the function that removes instructions activities from
`useRecentDatasetsActivitiesState` hook.

And use `addToRecentInstructionsActivities` method to register
a new activity ("deletion" activity) in `useInstructionsStatesManagement`
hook instead of using `removeRecentInstructionsActivities` function
(which has been deleted) to remove all related activities to the
deleted instruction.

Make `deleteInstructionState` function in `useInstructionsStatesManagement`
hook takes the full data of the deleted instruction to register the
activity with its data.

Make `DeleteInstructionButton` component takes the full
data of the dataset as prop to pass it to `useDeleteInstruction`
hook that passes the full dataset data to `deleteInstructionState`
function to update the local state after the dataset has been
deleted successfully.