feat: create `closeDialog` function in `DialogWrapper.tsx`
Use `useRef` hook to create a refrence to pass it to close button
inside `DialogContent` component,

And create `closeDialog` function to use the refrence to trigger
a click event on close button to close the dialog.

Make the children of `DialogContent` able to access `closeDialog`
function by passing to `content` prop a function receives as
parameter an object has `closeDialog` function.