fix: set a polyfil for downloading datasets via the app feature
Make "downloading datasets via the browser" feature works as polyfil
for "downloading datasets via the app" feature that depends on
`showSaveFilePicker` method that allows it to write data into a file in
users file system is not whidly supported in all browsers. So, Add an
additional condition to ensure that `showSaveFilePicker` method is
available before use it.

Also disable "App" option for the users who opened the application
via not supported browsers.